\name{WriteParquet}
\alias{WriteParquet}
\title{
WriteParquet
}
\description{
Writes Parquet files.
}
\usage{
WriteParquet(Data, FileName, FilePath = "")
}
\arguments{
\item{Data}{
List of matrices.
}
\item{FileName}{
Character Sting with name of file.
}
\item{FilePath}{
Character Sting with name of directory.
}
}
\author{
QS
}
\examples{

}
\keyword{parquet}
\keyword{DataIO}
